+++ Changelog.md
@@ -11,6 +11,7 @@ WhateverGreen Changelog
 - Added `applbkl-name` and `applbkl-data` IORegistry data keys to provide custom backlight data
 - Fixed applying CoreFP patches on Apple firmware, when they are not needed
 - Added `shikigva=16` (repurposed) property to use AMD hardware DRM decoder in select apps
+- Added `shikigva=128` (repurposed) property to use hardware decoder for FairPlay 1.0 (can be used as `shikigva=144`)
 
 #### v1.3.4
 - Added support for disabled AppleGraphicsDevicePolicy in AMD drivers on 10.15.1